Pre-CRN refactor cleanup (#272)
* Do the large line change code movement ahead of the CRN algorithm refactor

* Couple of doc fixes
