+++ docs/source/api_reference/framework/population/index.rst
@@ -1,3 +1,7 @@
+=====================
+Population Management
+=====================
+
 .. automodule:: vivarium.framework.population
 
 .. toctree::
