--- docs/source/api_reference/framework/randomness/core.rst
@@ -1,1 +1,0 @@
-.. automodule:: vivarium.framework.randomness.core
