+++ docs/source/api_reference/framework/randomness/index.rst
@@ -1,3 +1,7 @@
+========================
+Random Number Generation
+========================
+
 .. automodule:: vivarium.framework.randomness
 
 .. toctree::
