+++ src/vivarium/framework/randomness/__init__.py
@@ -21,11 +21,14 @@ These seeds are then used to generate `numpy.random.RandomState` objects that
 can be used to create pseudo-random numbers in a repeatable manner.
 
 
-For mor information, see the Common Random Numbers
+For more information, see the Common Random Numbers
 :ref:`concept note <crn_concept>`.
 
 """
-from vivarium.framework.randomness.core import RESIDUAL_CHOICE, choice, get_hash, random
 from vivarium.framework.randomness.exceptions import RandomnessError
 from vivarium.framework.randomness.manager import RandomnessInterface, RandomnessManager
-from vivarium.framework.randomness.stream import RandomnessStream
+from vivarium.framework.randomness.stream import (
+    RESIDUAL_CHOICE,
+    RandomnessStream,
+    get_hash,
+)
