+++ tests/framework/test_state_machine.py
@@ -2,7 +2,6 @@ import numpy as np
 import pandas as pd
 
 from vivarium import InteractiveContext
-from vivarium.framework.randomness import choice
 from vivarium.framework.state_machine import Machine, State, Transition
 
 
@@ -29,11 +28,12 @@ def _even_population_fixture(column, values):
             return "test_pop_fixture"
 
         def setup(self, builder):
+            self.randomness = builder.randomness.get_stream(self.name)
             self.population_view = builder.population.get_view([column])
             builder.population.initializes_simulants(self.inner, creates_columns=[column])
 
         def inner(self, pop_data):
-            self.population_view.update(choice("start", pop_data.index, values))
+            self.population_view.update(self.randomness.choice(pop_data.index, values))
 
     return pop_fixture()
